Changes of the pull request implemented

+++ Task4/src/main/java/ContainsW.java
@@ -9,7 +9,8 @@ public class ContainsW {
     public boolean containsw (String word) {
         char[] wordChars = word.toCharArray();
         // It can be an integer int numberOfW = 0;
-        byte numberOfW = 0;
+        // modified
+        int numberOfW = 0;
         for (char actualChar : wordChars) {
             if (actualChar == 'w') numberOfW++;
         }
